% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pronounCounter.R
\name{allPronounCount}
\alias{allPronounCount}
\title{Function which returns the count of all pronouns in the document}
\usage{
allPronounCount(wordFreqMatrix)
}
\arguments{
\item{wordFreqMatrix, }{the output from getFromFolderWF() or processXMLwf()}
}
\description{
This function returns the count of all the pronouns in the documents
}
\keyword{count,}
\keyword{frequency,}
\keyword{pronoun}
\keyword{token,}
\keyword{tokenize,}
\keyword{word}
