% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeTokenStatistics.R
\name{arthMean}
\alias{arthMean}
\title{Function to get the aritmetic mean}
\usage{
arthMean(typeTokenRatios)
}
\arguments{
\item{typeTokenRatios, }{the results from the type token ratio calculator}
}
\description{
This function takes the type token ratio data structure as an
argument and returns the arithmetic mean
}
\keyword{mean}
\keyword{ratios,}
\keyword{token}
\keyword{token,}
\keyword{type}
