% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleMaths.R
\name{findUnique}
\alias{findUnique}
\title{Function which finds tokens in the first element that do not exisit in the second element and returns them}
\usage{
findUnique(wordFreq1, names2)
}
\arguments{
\item{names2, }{the names of a vector which correspond to the tokens of another word frequency data structure (for efficency)}

\item{wordFreqA, }{a word frequency data structure that you want to find the unique tokens from}
}
\description{
Function which finds tokens that exsist in document B which do not exisit in document A
then adds them with a frequency of the smoothing factor to an updated list of tokens and 
frequencies for each document.  The relative frequency is then calculated
}
\keyword{find}
\keyword{unique}
\keyword{unique,}
