% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{getCount}
\alias{getCount}
\title{Function which returns the count of a word given the data frame
generated by the corpus level word frequency function getFromFolderWF
and the word requested}
\usage{
getCount(wordCountDataFrame, searchString)
}
\arguments{
\item{wordCountDataFrame, }{the unique tokens and their frequencies}

\item{searchString, }{the string you would like to know the frequency of}
}
\description{
Function which returns the count of a word given the data frame
generated by the corpus level word frequency function getFromFolderWF
and the word requested
}
\keyword{count}
\keyword{frequency,}
\keyword{get}
\keyword{token,}
\keyword{tokenize,}
\keyword{word}
