% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{getFixedSampleSizeS}
\alias{getFixedSampleSizeS}
\title{Function which returns a string that is the requested size}
\usage{
getFixedSampleSizeS(text, minMaxWordCount = 300)
}
\arguments{
\item{text, }{the text you would like tokenized}

\item{minMaxWordCount, }{no documents with less tokens than indicated will be accepted and all documents longer than the spefified count will be cropped Defaults to 300}
}
\description{
This function takes a string as an argument and returns null if the
string has less than 300 tokens or the first 300 tokens if 
it is 300 tokens or longer
}
\keyword{frequency,}
\keyword{sample}
\keyword{size}
\keyword{token,}
\keyword{tokenize,}
\keyword{word}
