% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{getFixedSampleSizeWF}
\alias{getFixedSampleSizeWF}
\title{function which returns a word frequency data frame from the portion
of a text which equal in length to the requested sample size}
\usage{
getFixedSampleSizeWF(text, minMaxWordCount = 300)
}
\arguments{
\item{text, }{the text you would like tokenized}

\item{minMaxWordCount, }{no documents with less tokens than indicated will be accepted and all documents longer than the spefified count will be cropped Defaults to 300}
}
\description{
This function takes a string as an argument and returns null if the
string has less than 300 tokens or a word frequency data frame
of the first 300 token if it is 300 tokens or longer
}
\keyword{frequency,}
\keyword{sample}
\keyword{size}
\keyword{token,}
\keyword{tokenize,}
\keyword{word}
