% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{getFromFolderString}
\alias{getFromFolderString}
\title{Function which takes a folder as an argument and returns a vector
with the unique tokens for each document and their frequencies}
\usage{
getFromFolderString(pathToFolder, minMaxWordCount = 300)
}
\arguments{
\item{pathToFolder, }{the folder containing your corpus}

\item{minMaxWordCount, }{no documents with less tokens than indicated will be accepted and all documents longer than the spefified count will be cropped Defaults to 300}
}
\description{
This function takes a path to a folder as an argument.  It then
excludes all documents in the folder which have less than 300 
tokens and process the rest.  The output is a vector of 
strings containing each acceptable, cropped document
}
\keyword{frequency,}
\keyword{sample}
\keyword{size}
\keyword{token,}
\keyword{tokenize,}
\keyword{word}
