% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicalTagger.R
\name{getLexTag}
\alias{getLexTag}
\title{Retrieves the tagged tokens from a csv file and places them in a
data structure}
\usage{
getLexTag(workingDirectory)
}
\arguments{
\item{workingDirectory, }{the folder where the tag/token csv files are stored}
}
\description{
This function goes into the directory specified and finds the output
from writeLexTag and puts it in a data structure
}
\keyword{lexical,}
\keyword{of}
\keyword{part}
\keyword{phrasemachine}
\keyword{speach,}
\keyword{tag,}
