% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{getOneDataFrame}
\alias{getOneDataFrame}
\title{Function which takes a path to a folder and returns one data frame}
\usage{
getOneDataFrame(pathToFolder, minMaxWordCount = 300)
}
\arguments{
\item{pathToFolder, }{the folder containing your corpus}

\item{minMaxWordCount, }{no documents with less tokens than indicated will be accepted and all documents longer than the spefified count will be cropped Defaults to 300}
}
\description{
Function which takes a path to a folder and returns
a single data frame of the tokens and their frequencies
}
\keyword{frequency,}
\keyword{sample}
\keyword{size}
\keyword{token,}
\keyword{tokenize,}
\keyword{word}
