% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicalTagger.R
\name{getTagCategoryCount}
\alias{getTagCategoryCount}
\title{Function to get the count of any given part of speach}
\usage{
getTagCategoryCount(taggedData, tagType, option = "all")
}
\arguments{
\item{taggedData, }{the output from the retieval function, getLexTag}

\item{tagType, }{the couple letter code from the list in the description}

\item{option, }{all or anything else will be accepted, all is the default and will return a cumulative total for the whole corpus, any other option will return individual counts for each document in a vector}
}
\description{
This function returns the count(s) of a given part of speech either as
a cumulative total (option = "all", the default) or a vector containing
the counts for each corresponing documents.  Acceptable tag types are as follows:
CC Coordinating conjunction, CD cardinal number, DT determiner
EX existential there, FW foreign word, IN preposition or
subordinating conjunction, JJ adjective, JJR Adjective, superlative
LS list item marker, MD modal, NN noun, singular or mass
NNS noun, plural, NNP proper noun, singular, NNPS proper noun, plural
PDT predeterminer, POS, possessive ending, PRP personal pronoun,
PRP$ possessive pronoun, RB adverb, RBR adverb, comparative
RBS adverb, superlative, RP particle, SYM symbol, TO to
UH interjection VB, verb, base form, VBD, verb, past tense
VBG, Verb gerund or present participle
VBP verb non 3rd person singular present
VBZ verb 3rd person singular present, WDT whdeterminer
WP whpronoun, WP$ possessive whpronoun, WRB, whadverb
}
\keyword{lexical,}
\keyword{of}
\keyword{part}
\keyword{phrasemachine}
\keyword{speach,}
\keyword{tag,}
