% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeTokenStatistics.R
\name{getTypeTokenRatios}
\alias{getTypeTokenRatios}
\alias{getTypeTokenRatios}
\title{Function to get the type token ratios for a corpus}
\usage{
getTypeTokenRatios(wordFrequencyMatrix)

getTypeTokenRatios(wordFrequencyMatrix)
}
\arguments{
\item{wordFrequencyMatrix, }{a data strucutre generated by this package which contians the unique tokens and their counts}

\item{path, }{the path to the folder containing the corpus}

\item{minMaxWordCount, }{no documents with less tokens than indicated will be accepted and all documents longer than the spefified count will be cropped, defaults to 300.}
}
\description{
This function cuts documents from a corpus to a fixed word count and
ignores the documents which are shorter than that word count.  The
accepted documents are then tokenized and the type/token ratios
for each are calculated and returned.

This function takes a data structure created by this package
using methods such as getFromFolderWF and returns the type token
ratios.
}
\keyword{ratios}
\keyword{ratios,}
\keyword{token}
\keyword{token,}
\keyword{type}
