% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeTokenStatistics.R
\name{interquartileRange}
\alias{interquartileRange}
\title{Function to calculate the interquartile range}
\usage{
interquartileRange(typeTokenRatios)
}
\arguments{
\item{typeTokenRatios, }{the results from the type token ratio calculator}
}
\description{
This function takes a data structure containing type token ratios
created by this package and returns the interquartile range
}
\keyword{interquartile}
\keyword{range}
\keyword{ratios,}
\keyword{token}
\keyword{token,}
\keyword{type}
