% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicalTagger.R
\name{lexTagFiles}
\alias{lexTagFiles}
\title{lexicalTaggging file function}
\usage{
lexTagFiles(pathToFolder, pathToStorage, minMaxWordCount = 300)
}
\arguments{
\item{pathToFolder, }{the path to the folder containing the corpus}

\item{pathToStorage, }{the path to the folder you would like the output to go}

\item{minMaxWordCount, }{no documents with less tokens than indicated will be accepted and all documents longer than the spefified count will be cropped Defaults to 300}
}
\description{
This function allows you to part of speech tag your corpus and write
the resulting tag/token relationships to a csv file.  This specific
function also uses a path to a folder containing *.txt documents
as input, as opposed to a pre-processed string
}
\keyword{lexical,}
\keyword{of}
\keyword{part}
\keyword{phrasemachine}
\keyword{speach,}
\keyword{tag,}
