% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicalTagger.R
\name{lexTagString}
\alias{lexTagString}
\title{lexicalTaggging string function}
\usage{
lexTagString(documents, pathToStorage)
}
\arguments{
\item{documents, }{a vector or list containing the strings to be tagged}

\item{pathToStorage, }{the path to the folder you would like the output to go}
}
\description{
This function allows you to part of speech tag your corpus and write
the resulting tag/token relationships to a csv file.  This specific
function uses a pre-processed vector of strings for input, as opposed
to gathering such a vector itself from a folder containing a corpus
}
\keyword{lexical,}
\keyword{of}
\keyword{part}
\keyword{phrasemachine}
\keyword{speach,}
\keyword{tag,}
