% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{make300}
\alias{make300}
\title{Function which cuts strings to the requested sample size
and returns null if they are shorter than the sample size}
\usage{
make300(text, minMaxWordCount)
}
\arguments{
\item{text, }{the text you would like tokenized}

\item{minMaxWordCount, }{no documents with less tokens than indicated will be accepted and all documents longer than the spefified count will be cropped Defaults to 300}
}
\description{
This function takes a string as an argument and returns null if the
string has less than 300 tokens or a word frequency data frame
of the first 300 token if it is 300 tokens or longer
}
\keyword{sample}
\keyword{size}
