% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeTokenStatistics.R
\name{medTypeToken}
\alias{medTypeToken}
\title{Function to calculate the median}
\usage{
medTypeToken(typeTokenRatios)
}
\arguments{
\item{typeTokenRatios, }{the results from the type token ratio calculator}
}
\description{
This function takes a data structure containing type token ratios
created by this package and returns the median
}
\keyword{median}
\keyword{ratios,}
\keyword{token}
\keyword{token,}
\keyword{type}
