% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleMaths.R
\name{mergeWF}
\alias{mergeWF}
\title{Function to join two lists of lists into one data frame}
\usage{
mergeWF(wordFreqOne, wordFreqTwo)
}
\arguments{
\item{wordFreqOne, }{the first list of lists to merge}

\item{wordFreqTwo, }{the second list of lists to merge}
}
\description{
Function which takes two lists of lists, converts them to data frames
and binds them together into one data frame
}
\examples{
mergeWF(wordFreqA, uniqueFreq)
}
\keyword{data}
\keyword{frame}
\keyword{list}
\keyword{matrix,}
\keyword{merge,}
