% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeTokenStatistics.R
\name{mode}
\alias{mode}
\title{Function to calculate the mode}
\usage{
mode(typeTokenRatios)
}
\arguments{
\item{typeTokenRatios, }{the results from the type token ratio calculator}
}
\description{
This function takes a data structure containing type token ratios
created by this package and returns the mode
}
\keyword{mode}
\keyword{ratios,}
\keyword{token}
\keyword{token,}
\keyword{type}
