% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeTokenStatistics.R
\name{plotDist}
\alias{plotDist}
\title{Function to plot the normal distribution and actual distribution}
\usage{
plotDist(typeTokenRatios)
}
\arguments{
\item{typeTokenRatios, }{the results from the type token ratio calculator}
}
\description{
This function takes a data structure containing type token ratios
created by this package plots the standard and actual definition
}
\keyword{deviation}
\keyword{ratios,}
\keyword{standard}
\keyword{token}
\keyword{token,}
\keyword{type}
