% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeTokenStatistics.R
\name{printMeasuresOfCentralTendency}
\alias{printMeasuresOfCentralTendency}
\title{Function to print the measures of central tendency}
\usage{
printMeasuresOfCentralTendency(typeTokenRatios)
}
\arguments{
\item{typeTokenRatios, }{the results from the type token ratio calculator}
}
\description{
This function takes a data structure containing type token ratios
created by this package and prints the measures central tendency
}
\keyword{central}
\keyword{ratios,}
\keyword{tendency}
\keyword{token}
\keyword{token,}
\keyword{type}
