% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeTokenStatistics.R
\name{printMeasuresOfDispersion}
\alias{printMeasuresOfDispersion}
\title{Function to print the measures of dispersion}
\usage{
printMeasuresOfDispersion(typeTokenRatios)
}
\arguments{
\item{typeTokenRatios, }{the results from the type token ratio calculator}
}
\description{
This function takes a data structure containing type token ratios
created by this package and prints the measures of dispersion
}
\keyword{dispersion}
\keyword{ratios,}
\keyword{token}
\keyword{token,}
\keyword{type}
