% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlRead.R
\name{processXMLstring}
\alias{processXMLstring}
\title{Function to process XML data from a specific corpus into a vector of strings}
\usage{
processXMLstring(pathToFolder, minMaxxWordCount = 300)
}
\arguments{
\item{pathToFolder, }{the path to the folder containing the corpus}

\item{minMaxWordCount, }{no documents with less tokens than indicated will be accepted and all documents longer than the spefified count will be cropped Defaults to 300}
}
\description{
This function takes a .xml documents from a corpus of forum posts
and returns a vector of strings.  Can perhaps
be used for other forum corpora which have a similar structure
}
\keyword{cleaning,}
\keyword{corpus}
\keyword{processing,}
\keyword{xml,}
