% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlRead.R
\name{processXMLwithTrainingLabels}
\alias{processXMLwithTrainingLabels}
\title{Function for coping and sorting files}
\usage{
processXMLwithTrainingLabels(pathToTrainingLabels, pathToTrainingData,
  destinationFolder, type)
}
\arguments{
\item{pathToTrainingLabels, }{path to the tab separtated value file}

\item{pathToTrainingData, }{path to the data which is described by the tsv file}

\item{destinationFolder, }{folder where the files will be copied,}

\item{type, }{the label used in the tsv to describe the requested files it is also assumed to be the name of the folder inside the requested directory where the files will go}
}
\description{
which takes a tab seperated value file containing the ids of files
which have already been tagged as belonging to a given category and
copies them into a specified folder so they can be processed seperately
}
\keyword{cleaning,}
\keyword{corpus}
\keyword{processing,}
\keyword{xml,}
