% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pronounCounter.R
\name{pronounCount}
\alias{pronounCount}
\title{Function which returns the count of a specific kind of pronoun}
\usage{
pronounCount(wordFreqMatix, pronounType)
}
\arguments{
\item{pronounType, }{a number in [1,5] which maps to a list of pronouns in a manner described in the function description}

\item{wordFreqMatrix, }{the output from getFromFolderWF() or processXMLwf()}
}
\description{
This function returns the count of a specific kind of pronoun. 
The kind of pronouns in use can be specified using the following 
numbers:
1 = first person singular pronoun (fpsp)
2 = first person plural pronoun (fppp)
3 = third person singular pronoun (tpsp)
4 = third person plural pronoun (tppp)
5 = second person pronouns (spp)
}
\keyword{count,}
\keyword{frequency,}
\keyword{pronoun}
\keyword{token,}
\keyword{tokenize,}
\keyword{word}
