% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeTokenStatistics.R
\name{rangeTypeToken}
\alias{rangeTypeToken}
\title{Function to calculate the range}
\usage{
rangeTypeToken(typeTokenRatios)
}
\arguments{
\item{typeTokenRatios, }{the results from the type token ratio calculator}
}
\description{
This function takes a data structure containing type token ratios
created by this package and returns the range
}
\keyword{range}
\keyword{ratios,}
\keyword{token}
\keyword{token,}
\keyword{type}
