% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleMaths.R
\name{relativeFrequencies}
\alias{relativeFrequencies}
\title{Function to calculate the relative frequencies, given two word frequency data frames}
\usage{
relativeFrequencies(newWordFreqA, newWordFreqB)
}
\arguments{
\item{newWordFreqA, }{the target word frequency data frame}

\item{newWordFreqB, }{the reference word frequency data fram}
}
\description{
Function which, given two word frequency data frames generated by
frequencies() and the mergeWF() function, returns a data frame
with the tokens and their relative frequencies
}
\examples{
relativeFrequencies(newWordFrequenciesA, newWordFrequenciesB)
}
\keyword{data}
\keyword{frame}
\keyword{list}
\keyword{matrix,}
\keyword{merge,}
