% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleMaths.R
\name{simpleMaths}
\alias{simpleMaths}
\title{function to find the relative frequencies of the combined tokens from document A and document B}
\usage{
simpleMaths(wordFrequenciesA, wordFrequenciesB, smoothingFactor)
}
\arguments{
\item{wordFrequenciesA, }{a wordFrequency list, generated by this package}

\item{wordFrequenciesB, }{another word frequency list for reference}

\item{smoothingFactor, }{the number used when the frequency of a word is zero and also the smooting factor}
}
\description{
Function which finds tokens that exsist in document B which do not exisit in document A
then adds them with a frequency of the smoothing factor to an updated list of tokens and 
frequencies for each document.  The relative frequency is then calculated
}
\examples{
frequencies(wordFreqA, wordFreqB, 1000)
}
\keyword{frequency}
\keyword{relative}
\keyword{simplemaths,}
