% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeTokenStatistics.R
\name{standardDev}
\alias{standardDev}
\title{Function to calculate the standard deviation}
\usage{
standardDev(typeTokenRatios)
}
\arguments{
\item{typeTokenRatios, }{the results from the type token ratio calculator}
}
\description{
This function takes a data structure containing type token ratios
created by this package and returns the standard deviation
}
\keyword{deviation}
\keyword{ratios,}
\keyword{standard}
\keyword{token}
\keyword{token,}
\keyword{type}
