% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeTokenStatistics.R
\name{variance}
\alias{variance}
\title{Function to calculate the variance}
\usage{
variance(typeTokenRatios)
}
\arguments{
\item{typeTokenRatios, }{the results from the type token ratio calculator}
}
\description{
This function takes a data structure containing type token ratios
created by this package and returns the variance
}
\keyword{ratios,}
\keyword{token}
\keyword{token,}
\keyword{type}
\keyword{variance}
