% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlRead.R
\name{XMLgetOneDataFrame}
\alias{XMLgetOneDataFrame}
\title{Function to process XML data from a specific corpus into unique tokens and their frequencies}
\usage{
XMLgetOneDataFrame(pathToFolder, minMaxWordCount = 300)
}
\arguments{
\item{pathToFolder, }{the path to the folder containing the corpus}

\item{minMaxWordCount, }{no documents with less tokens than indicated will be accepted and all documents longer than the spefified count will be cropped, defaults to 300.  If it is set to NULL documents will not be cut and will only be excluded if they contain no usable characters}
}
\description{
This function takes a .xml documents from a corpus of forum posts
and returns the unique tokens and their frequencies in one data frame.  Can perhaps
be used for other forum corpora which have a similar structure
}
\keyword{cleaning,}
\keyword{corpus}
\keyword{processing,}
\keyword{xml,}
